Photo list tag and preview working
