+++ src/Patient/Progress/PhotoList/PhotoResponseItem.js
@@ -5,10 +5,8 @@ import { DateTime } from 'luxon';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
-import Tag from '../../../Components/Tag';
-import { useTranslation } from 'react-i18next';
 import ExpandablePhoto from '../../../Components/ExpandablePhoto';
-
+import PhotoResultTag from '../../../Components/Shared/PhotoResultTag';
 
 const useStyles = makeStyles({
     item: {
@@ -23,13 +21,6 @@ const useStyles = makeStyles({
         color: Colors.textDarkGray,
         lineHeight: "1em"
     },
-    tag: {
-        display: "block",
-        fontSize: ".8em",
-        width: "fit-content",
-        padding: "2px 8px",
-        fontWeight: "bold"
-    },
     reportData: {
         flexGrow: "1",
         display: "flex",
@@ -40,31 +31,6 @@ const useStyles = makeStyles({
         color: Colors.textDarkGray
     }
 })
-
-const Result = ({ approved, skipped }) => {
-
-    const { t } = useTranslation('translation');
-    const classes = useStyles();
-
-    let text = "";
-    let color = "";
-
-    if (skipped) {
-        color = Colors.calendarRed;
-        text = t('dashboard.skippedPhoto');
-    } else if (approved === null) {
-        color = Colors.lightgray;
-        text = t('photoReportReview.awaiting');
-    } else if (approved) {
-        color = Colors.calendarGreen;
-        text = t('photoReportReview.detected');
-    } else {
-        color = Colors.highlightYellow;
-        text = t('photoReportReview.unclear');
-    }
-
-    return <Tag className={classes.tag} backgroundColor={color}>{text}</Tag>
-}
 
 const PhotoResponseItem = ({ date, approved, url, whyPhotoWasSkipped }) => {
 
@@ -80,7 +46,7 @@ const PhotoResponseItem = ({ date, approved, url, whyPhotoWasSkipped }) => {
                 <div className={classes.reportData}>
                     <Typography className={classes.date}>{displayDate}</Typography>
                     <Box height=".33em" />
-                    <Result skipped={!url} approved={approved} />
+                    <PhotoResultTag skipped={!url} approved={approved} />
                     {!url && <>
                         <Box height=".33em" />
                         <Typography className={classes.reason} variant="body1">{whyPhotoWasSkipped || t('coordinator.sideBar.noReason')}</Typography>
